Modified Dates: Fixes ratification and draft dates
Since we technically just amended the constitution, changes the
ratified day to today. Also, we are supposed to have the "date
drafted" at the header, so that text has been updated as well as
the inclusion of the `today` LaTeX command so that CI takes care
of those dates on it's own.

Signed-off-by: Mark Stenglein <mark@stengle.in>